requirements.txt
Django==2.2.7
beautifulsoup4
bs4
certifi
chardet
dj-database-url
idna
lxml
pytz
requests
urllib3
whitenoise
Jinja2
MarkupSafe
psycopg2
django-allauth
django-import-export
pillow
django-bootstrap4
